Ensure index is an integer - python 3
- When calculating core_start, python 2 returned integers when dividing integers. Python 3 returns a float so we convert it to an integer for use as an index

+++ predict_tf_binding.py
@@ -2,7 +2,7 @@
 
 import argparse
 import itertools
-from svmutil import *
+from libsvm.svmutil import *
 from math import exp
 from multiprocessing import Process
 
@@ -87,7 +87,7 @@ def generate_matching_sequences(sequence, core, width, core_start=None):
     # The core positions are calculated relative to the window (and not the overall sequence)
     # This works as long as both core and width are same parity
     if core_start is None:
-        core_start = (width - core_width) / 2
+        core_start = int((width - core_width) / 2)
     for start in range(max_start + 1):
         end = start + width
         window_sequence = sequence[start:end]
